+++ libcxx/test/std/utilities/memory/util.smartptr/util.smartptr.shared/util.smartptr.shared.create/make_shared.array.unbounded.pass.cpp
@@ -11,9 +11,6 @@
 // This test requires support for aligned allocation to test overaligned types.
 // XFAIL: availability-aligned_allocation-missing
 
-// https://reviews.llvm.org/D129198 is not in AppleClang 14
-// XFAIL: stdlib=apple-libc++ && target={{.+}}-apple-macosx10.13{{(.0)?}} && apple-clang-14
-
 // <memory>
 
 // shared_ptr